% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-S3.R
\name{as.data.frame.LEV_summary}
\alias{as.data.frame.LEV_summary}
\title{S3 method for extracting LEV Profiles Summaries}
\usage{
\method{as.data.frame}{LEV_summary}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{LEV_summary}}

\item{row.names}{Not used}

\item{optional}{Not used}

\item{...}{Extra arguments}
}
\value{
Data Frame
}
\description{
S3 method for extracting LEV Profiles Summaries
}
\examples{
# Create random athletes/profiles
sets <- create_athletes(1) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(70, 80, 90)) \%>\%
  create_summary()

sets_params <- as.data.frame(sets)
}
