% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-S3.R
\name{coef.LEV_profile}
\alias{coef.LEV_profile}
\title{S3 method for extracting LEV Profiles and Visits parameters}
\usage{
\method{coef}{LEV_profile}(object, ...)
}
\arguments{
\item{object}{Object of class \code{LEV_profile}}

\item{...}{Extra arguments}
}
\value{
Data Frame
}
\description{
S3 method for extracting LEV Profiles and Visits parameters
}
\examples{
profiles <- create_athletes(10)
profile_params <- coef(profiles)

visits <- profiles \%>\% create_visits(1:3)
visit_params <- coef(visits)
}
