% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-S3.R
\name{coef.LEV_sets}
\alias{coef.LEV_sets}
\title{S3 method for extracting LEV Profiles Sets}
\usage{
\method{coef}{LEV_sets}(object, ...)
}
\arguments{
\item{object}{Object of class \code{LEV_sets}}

\item{...}{Extra arguments}
}
\value{
Data Frame
}
\description{
S3 method for extracting LEV Profiles Sets
}
\examples{
# Create random athletes/profiles
sets <- create_athletes(1) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(70, 80, 90))

sets_params <- coef(sets)
}
