% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-S3.R
\name{coef.LEV_summary}
\alias{coef.LEV_summary}
\title{S3 method for extracting LEV Profiles Summaries}
\usage{
\method{coef}{LEV_summary}(object, ...)
}
\arguments{
\item{object}{Object of class \code{LEV_summary}}

\item{...}{Extra arguments}
}
\value{
Data Frame
}
\description{
S3 method for extracting LEV Profiles Summaries
}
\examples{
# Create random athletes/profiles
sets <- create_athletes(1) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(70, 80, 90)) \%>\%
  create_summary()

sets_params <- coef(sets)
}
