% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-prescription-1RM.R
\name{create_prescription_1RM}
\alias{create_prescription_1RM}
\title{Create prescription 1RM}
\usage{
create_prescription_1RM(LEV_profile, init_1RM, change_1RM)
}
\arguments{
\item{LEV_profile}{Object returned from \code{\link{create_visits}}}

\item{init_1RM}{Initial 1RM value}

\item{change_1RM}{1RM change per visit}
}
\description{
This function creates prescription 1RM within provided \code{LEV_profile}
}
\examples{
set.seed(1667)
sets <- create_profiles(1, load_increment = 1) \%>\%
  create_visits(1:10) \%>\%
  create_prescription_1RM(100, 2.5) \%>\%
  create_sets(load = 0.8, load_type = "prescription 1RM")

plot(sets, x_var = "visit", y_var = "load", reps = 1, facet = NULL)

plot(sets, x_var = "visit", y_var = "prescription_1RM", facet = NULL)
}
