% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-summary.R
\name{create_summary}
\alias{create_summary}
\title{Create Summary}
\usage{
create_summary(LEV_sets)
}
\arguments{
\item{LEV_sets}{\code{LEV_sets} object, returned by \code{\link{create_sets}} function}
}
\value{
\code{LEV_summary} object
}
\description{
Function creates summary of the sets, using provided \code{LEV_sets}
}
\examples{
set.seed(1667)
sets <- create_athletes(2) \%>\%
  create_visits(1) \%>\%
  create_visit_1RM() \%>\%
  create_sets(load = c(0.7, 0.8, 0.9), load_type = "visit 1RM") \%>\%
  create_summary()

sets
plot(sets)
}
