% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-visit-1RM.R
\name{create_visit_1RM}
\alias{create_visit_1RM}
\title{Create visit 1RM}
\usage{
create_visit_1RM(
  LEV_profile,
  load_perc = seq(0.8, 1.2, by = 0.025),
  use_true_velocity = FALSE
)
}
\arguments{
\item{LEV_profile}{Object returned from \code{\link{create_visits}}}

\item{load_perc}{Percentage to be used to calculate loads using theoretical visit profile 1RM}

\item{use_true_velocity}{When estimating failure, should true or biological (default) velocity be used?}
}
\description{
This function creates visit 1RM within provided \code{LEV_profile}.
It simply tests the progressive load until one cannot make a single rep.
This is done by using \code{load_perc} against theoretical visit profile 1RM
}
\examples{
set.seed(1667)
sets <- create_profiles(1, L0_visit_change_additive = 2.5) \%>\%
  create_visits(1:10) \%>\%
  create_visit_1RM() \%>\%
  create_sets(load = 0.8, load_type = "visit 1RM")

plot(sets, x_var = "visit", y_var = "load", reps = 1, facet = NULL)

plot(sets, x_var = "visit", y_var = "visit_1RM", facet = NULL)
}
