% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-visits.R
\name{create_visits}
\alias{create_visits}
\title{Create Visits}
\usage{
create_visits(LEV_profile = create_profiles(), visit = 1)
}
\arguments{
\item{LEV_profile}{Object returned from \code{\link{create_profiles}} or \code{\link{create_athletes}}}

\item{visit}{Numeric vector. Default is 1}
}
\value{
\code{LEV_profile} object
}
\description{
Before sets can be generated, it is necessary to create visits to the laboratory/gym
}
\examples{
# Create random athletes/profiles
set.seed(1667)

sets <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(90, 110, 130),
    load_type = "absolute"
  )

print(sets)

# Extract data frame
LEV_data <- as.data.frame(sets)
# Or
# LEV_data <- coef(sets)

plot(sets)
plot(sets, athletes = "Athlete 1")
plot(sets, athletes = "Athlete 1", reps = 1)
plot(sets, athletes = "Athlete 1", x_var = "RIR")

# Another way to create LEV profiles
sets <- create_profiles(athletes = c("Mladen", "Ivan"), L0 = c(200, 180)) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(100, 120, 140), load_type = "absolute")

plot(sets)
plot(sets, facet = NULL, x_var = "load")
plot(sets, visits = 1, x_var = "RIR")
}
