% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-EVP.R
\name{estimate_EVP}
\alias{estimate_EVP}
\title{Estimate Exertion-Velocity Profile}
\usage{
estimate_EVP(
  rep,
  load,
  velocity,
  V0 = 1.8,
  L0 = 180,
  v1RM = 0.2,
  L0_rep_drop_limits = c(0, 0.2),
  V0_rep_drop_limits = c(0, 0.2),
  model_weights = 1,
  relative_load = FALSE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{rep}{Numeric vector indicating repetition}

\item{load}{Numeric vector indicating absolute load used. If relative load is used,
then use \code{relative_load = TRUE}}

\item{velocity}{Numeric vector indicating repetition velocity}

\item{V0}{Profile parameter}

\item{L0}{Profile parameter}

\item{v1RM}{Profile parameter. Used only if \code{relative_load = TRUE}}

\item{L0_rep_drop_limits}{Two element search vector, default is \code{c(0, 0.2)}}

\item{V0_rep_drop_limits}{Two element search vector, default is \code{c(0, 0.2)}}

\item{model_weights}{Observation weights. Default is \code{1}, implying
equal importance of the observations}

\item{relative_load}{If \code{load} is relative, then absolute load is calculated by
multiplying with profile 1RM}

\item{na.rm}{Should missing values be removed? Defaults is \code{FALSE}}

\item{...}{forwarded to \link[stats]{optim}}
}
\value{
Object \code{LEV_estimate}
}
\description{
Finds the \code{L0_reps_drop} and \code{V0_reps_drop} parameters that
minimize the sum-of-squares-error given the performed \code{rep}, \code{load}, and \code{velocity}
and \code{V0} and \code{L0} parameters.
}
\examples{
RTF_df <- create_profiles(biological_variation_multiplicative = 0.05, L0_rep_drop = 0.06) \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "RTF")

m1 <- estimate_EVP(
  rep = RTF_df$rep,
  load = RTF_df$load,
  velocity = RTF_df$measured_rep_velocity
)

m1
}
