% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-LVP.R
\name{estimate_LVP}
\alias{estimate_LVP}
\title{Estimate Load-Velocity Profile}
\usage{
estimate_LVP(load, velocity, model_weights = load, na.rm = FALSE)
}
\arguments{
\item{load}{Numeric vector indicating load used}

\item{velocity}{Numeric vector indicating repetition velocity}

\item{model_weights}{Observation weights. Default is \code{load}, implying
the higher the load, the higher the importance of the observation}

\item{na.rm}{Should missing values be removed? Defaults is \code{FALSE}}
}
\value{
Object \code{LEV_estimate}
}
\description{
Finds the \code{L0} and \code{V0} parameters using simple linear
regression given the \code{load} and \code{velocity} performance
}
\examples{
LV_df <- create_profiles() \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "LV")

m1 <- estimate_LVP(LV_df$load, LV_df$measured_rep_velocity)

m1
}
