% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-RTF.R
\name{estimate_RTF}
\alias{estimate_RTF}
\title{Estimate Reps to Failure Profile using Optim and Fractional MNR}
\usage{
estimate_RTF(
  load,
  MNR,
  V0 = 1.8,
  L0 = 180,
  v1RM = 0.2,
  L0_rep_drop_limits = c(0, 0.2),
  V0_rep_drop_limits = c(0, 0.2),
  model_weights = 1,
  relative_load = FALSE,
  max_reps = 100,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{load}{Numeric vector indicating absolute load used. If relative load is used,
then use \code{relative_load = TRUE}}

\item{MNR}{Numeric vector indicating maximum number of repetitions performed}

\item{V0}{Profile parameter}

\item{L0}{Profile parameter}

\item{v1RM}{Profile parameter}

\item{L0_rep_drop_limits}{Two element search vector, default is \code{c(0, 0.2)}}

\item{V0_rep_drop_limits}{Two element search vector, default is \code{c(0, 0.2)}}

\item{model_weights}{Observation weights. Default is \code{1}, implying
equal importance of the observations}

\item{relative_load}{If \code{load} is relative, then absolute load is calculated by
multiplying with profile 1RM}

\item{max_reps}{How many maximum reps to generate to search for failure? Default is 100}

\item{na.rm}{Should missing values be removed? Defaults is \code{FALSE}}

\item{...}{forwarded to \link[stats]{optim}}
}
\value{
Object \code{LEV_estimate}
}
\description{
Finds the \code{L0_reps_drop} and \code{V0_reps_drop} parameters using the \link[stats]{optim}
function that minimizes the sum-of-squares-error given the performed \code{load} and \code{MNR} and \code{V0},
\code{L0}, and \code{v1RM} parameters. Fractional MNRs are utilized in this function compare to
the \code{\link{estimate_RTF_grid}}
}
\examples{
RTF_df <- create_profiles() \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  create_summary() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "RTF")

m1 <- estimate_RTF(
  RTF_df$load,
  RTF_df$nRM
)

m1
}
