% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-perc-1RM.R
\name{get_perc_1RM}
\alias{get_perc_1RM}
\title{Get Maximum Percent of 1RM}
\usage{
get_perc_1RM(
  MNR,
  V0 = 1.8,
  L0 = 180,
  v1RM = 0.2,
  L0_rep_drop = 0.045,
  V0_rep_drop = 0,
  max_reps = 100,
  interval = c(0.5, 1.1),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{MNR}{Maximum number of repetitions}

\item{V0}{Profile parameter}

\item{L0}{Profile parameter}

\item{v1RM}{Profile parameter}

\item{L0_rep_drop}{Profile parameter}

\item{V0_rep_drop}{Profile parameter}

\item{max_reps}{How many maximum reps to generate to search for failure? Default is 100}

\item{interval}{Two element search vector. Default is \code{c(0.5, 1.1)}}

\item{na.rm}{Should missing values be removed? Defaults is \code{FALSE}}

\item{...}{Forwarded to \code{\link[stats]{uniroot}} function}
}
\value{
Numeric vector
}
\description{
Find the maximum percent of 1RM given the
\code{MNR}, \code{L0}, \code{V0}, \code{v1RM}, \code{L0_rep_drop}, and \code{V0_rep_drop}
parameters.
}
\examples{
get_perc_1RM(c(1, 2, 3, 4, 5))
}
