% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-model-fit.R
\name{LEV_model_fit}
\alias{LEV_model_fit}
\title{Function to return model fit metrics}
\usage{
LEV_model_fit(model = NULL, observed, predicted, na.rm = FALSE)
}
\description{
Function to return model fit metrics
}
