% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-S3.R
\name{plot.LEV_sets}
\alias{plot.LEV_sets}
\title{S3 method for plotting LEV Profiles Sets}
\usage{
\method{plot}{LEV_sets}(x, type = "athletes", ...)
}
\arguments{
\item{x}{Object of class \code{LEV_sets}}

\item{type}{Type of plot. Default is "athletes" which creates facets
for each athlete. Another option is "pooled", which doesn't create
facets for athletes by default. One can use \code{facet} parameter
to create facets using another variable. See Details}

\item{...}{Extra arguments. See Details}
}
\description{
S3 method for plotting LEV Profiles Sets
}
\details{
S3 method for plotting LEV Profiles Sets provides plenty of
flexibility and control by using the \code{...} parameters.
User can filter variables, use different x and y scales,
provides colors and facets. Here are the extra parameters your
can utilize:
\describe{
\item{athletes}{Althetes to keep in}
\item{visits}{Visits to keep in}
\item{sets}{Sets to keep in}
\item{loads}{Loads to keep in}
\item{RTFs}{Reps-To-Failure indicator (TRUE/FALSE) to keep in}
\item{reps}{Repetitions to keep in}
\item{RIRs}{Reps-In-Reserve to keep in}
\item{nRMs}{N-Repetition-Maximum loads to keep in}
\item{failed_reps}{Should failed reps be kept in? Default is \code{FALSE}}
\item{y_var}{Variable on y-axis. Default is 'measured_rep_velocity'}
\item{x_var}{Variable on x-axis. Default is 'load'}
\item{smooth}{Should smoothing be use? Default is \code{TRUE}}
\item{method}{Method for smoothing. Default is 'lm'. See
\code{\link[ggplot2]{geom_smooth}} for more information}
\item{formula}{Equation formula for smoothing. Default is \code{y~x}}
\item{se}{Should confidence interval be provided? Default is \code{TRUE}}
\item{level}{Level of confidence. Default is 0.9 (i.e., 90\%)}
\item{color}{What variable should be used to coloring groups? Default is none}
\item{facet}{Variable to be used for plot faceting. Athletes are already faceted,
and this parameter allows for adding extra level. Default is 'visit'}
\item{fullrange}{Should smooth/regression line be extended? Default is \code{FALSE}}
}

Please check the examples section for few examples
}
\examples{
# Create random athletes/profiles
set.seed(1667)

sets <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(90, 110, 130),
    load_type = "absolute"
  )

print(sets)

# Extract data frame
LEV_data <- as.data.frame(sets)
# Or
# LEV_data <- coef(sets)

plot(sets)
plot(sets, athletes = "Athlete 1")
plot(sets, athletes = "Athlete 1", reps = 1)
plot(sets, athletes = "Athlete 1", x_var = "RIR")

# Another way to create LEV profiles
sets <- create_profiles(athletes = c("Mladen", "Ivan"), L0 = c(200, 180)) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(100, 120, 140), load_type = "absolute")

plot(sets)
plot(sets, facet = NULL, x_var = "load")
plot(sets, visits = 1, x_var = "RIR")
}
