% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-S3.R
\name{plot.LEV_summary}
\alias{plot.LEV_summary}
\title{S3 method for plotting LEV Sets Summaries}
\usage{
\method{plot}{LEV_summary}(x, type = "athletes", ...)
}
\arguments{
\item{x}{Object of class \code{LEV_sets}}

\item{type}{Type of plot. Default is "athletes" which creates facets
for each athlete. Another option is "pooled", which doesn't create
facets for athletes by default. One can use \code{facet} parameter
to create facets using another variable. See Details}

\item{...}{Extra arguments. See Details}
}
\description{
S3 method for plotting LEV Sets Summaries
}
\details{
S3 method for plotting LEV Sets Summaries provides plenty of
flexibility and control by using the \code{...} parameters.
User can filter variables, use different x and y scales,
provides colors and facets. Here are the extra parameters your
can utilize:
\describe{
\item{athletes}{Althetes to keep in}
\item{visits}{Visits to keep in}
\item{loads}{Loads to keep in}
\item{RTFs}{Reps-To-Failure indicator (TRUE/FALSE) to keep in}
\item{reps}{Repetitions to keep in}
\item{RIRs}{Reps-In-Reserve to keep in}
\item{y_var}{Variable on y-axis. Default is 'measured_rep_velocity'}
\item{x_var}{Variable on x-axis. Default is 'load'}
\item{color}{What variable should be used to coloring groups? Default is none}
\item{facet}{Variable to be used for plot faceting. Athletes are already faceted,
and this parameter allows for adding extra level. Default is 'visit'}
}

Please check the examples section for few examples
}
\examples{
set.seed(1667)
sets <- create_athletes(2) \%>\%
  create_visits(1) \%>\%
  create_visit_1RM() \%>\%
  create_sets(load = c(0.7, 0.8, 0.9), load_type = "visit 1RM") \%>\%
  create_summary()

sets
plot(sets)
}
