% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-S3.R
\name{+.LEV_profile}
\alias{+.LEV_profile}
\title{S3 method for adding two or more LEV Profiles and Visits}
\usage{
\method{+}{LEV_profile}(lhs, rhs)
}
\arguments{
\item{lhs}{Object of class \code{LEV_profile}}

\item{rhs}{Object of class \code{LEV_profile}}
}
\value{
Object of class \code{LEV_profile}
}
\description{
S3 method for adding two or more LEV Profiles and Visits
}
\examples{
# Add profiles
profile1 <- create_athletes(5)
profile2 <- create_profiles(athletes = c("Mladen", "Milan", "Mihailo"))

added_profiles <- profile1 + profile2
added_profiles

# Add visits
visits1 <- profile1 \%>\%
  create_visits(1:3)

visits2 <- profile2 \%>\%
  create_visits(1:10)

added_visits <- visits1 + visits2
added_visits
}
