% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-S3.R
\name{+.LEV_sets}
\alias{+.LEV_sets}
\title{S3 method for adding two or more LEV Sets}
\usage{
\method{+}{LEV_sets}(lhs, rhs)
}
\arguments{
\item{lhs}{Object of class \code{LEV_sets}}

\item{rhs}{Object of class \code{LEV_sets}}
}
\value{
Object of class \code{LEV_sets}
}
\description{
S3 method for adding two or more LEV Sets
}
\examples{
sets1 <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(70, 80, 90)
  )

sets2 <- create_profiles(athletes = c("Mladen", "Milan", "Mihailo")) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(70, 80, 90)
  )

add_sets <- sets1 + sets2

add_sets
}
