% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-S3.R
\name{+.LEV_summary}
\alias{+.LEV_summary}
\title{S3 method for adding two or more LEV Summaries}
\usage{
\method{+}{LEV_summary}(lhs, rhs)
}
\arguments{
\item{lhs}{Object of class \code{LEV_sets}}

\item{rhs}{Object of class \code{LEV_sets}}
}
\value{
Object of class \code{LEV_sets}
}
\description{
S3 method for adding two or more LEV Summaries
}
\examples{
sets1 <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(70, 80, 90)
  ) \%>\%
  create_summary()

sets2 <- create_profiles(athlete = c("Mladen", "Milan", "Mihailo")) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(70, 80, 90)
  ) \%>\%
  create_summary()

add_sets <- sets1 + sets2

add_sets
}
