% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-S3.R
\name{predict.LEV_estimate}
\alias{predict.LEV_estimate}
\title{S3 method for making predictions using \code{LEV_estimate}}
\usage{
\method{predict}{LEV_estimate}(object, ...)
}
\arguments{
\item{object}{\code{LEV_estimate} object}

\item{...}{Forwarded to generic \code{predict()} function}
}
\description{
S3 method for making predictions using \code{LEV_estimate}
}
\examples{
LV_df <- create_profiles() \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "LV")

m1 <- estimate_LVP(LV_df$load, LV_df$measured_rep_velocity)

predict(m1, newdata = data.frame(load = 100))
}
