% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-S3.R
\name{print.LEV_estimate}
\alias{print.LEV_estimate}
\title{S3 method for printing \code{LEV_estimate} object}
\usage{
\method{print}{LEV_estimate}(x, ...)
}
\arguments{
\item{x}{\code{LEV_estimate} object}

\item{...}{Not used}
}
\description{
S3 method for printing \code{LEV_estimate} object
}
\examples{
LV_df <- create_profiles() \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "LV")

m1 <- estimate_LVP(LV_df$load, LV_df$measured_rep_velocity)

m1
}
