% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-S3.R
\name{print.LEV_profile}
\alias{print.LEV_profile}
\title{S3 method for printing LEV Profiles and Visits}
\usage{
\method{print}{LEV_profile}(x, ...)
}
\arguments{
\item{x}{Object of class \code{LEV_profile}}

\item{...}{Extra arguments}
}
\description{
S3 method for printing LEV Profiles and Visits
}
\examples{
profiles <- create_athletes(10)
print(profiles)

visits <- profiles \%>\% create_visits(1:3)
print(visits)
}
