% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-S3.R
\name{print.LEV_sets}
\alias{print.LEV_sets}
\title{S3 method for printing LEV Profiles Sets}
\usage{
\method{print}{LEV_sets}(x, ...)
}
\arguments{
\item{x}{Object of class \code{LEV_sets}}

\item{...}{Extra arguments}
}
\description{
S3 method for printing LEV Profiles Sets
}
\examples{
# Create random athletes/profiles
sets <- create_athletes(10) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(70, 80, 90))

print(sets)
}
