% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-S3.R
\name{residuals.LEV_estimate}
\alias{residuals.LEV_estimate}
\title{S3 method for providing residuals for the \code{LEV_estimate} object}
\usage{
\method{residuals}{LEV_estimate}(object, ...)
}
\arguments{
\item{object}{\code{LEV_estimate} object}

\item{...}{Not used}
}
\description{
S3 method for providing residuals for the \code{LEV_estimate} object
}
\examples{
LV_df <- create_profiles() \%>\%
  create_visits() \%>\%
  create_tests() \%>\%
  as.data.frame() \%>\%
  dplyr::filter(set == "LV")

m1 <- estimate_LVP(LV_df$load, LV_df$measured_rep_velocity)

residuals(m1)
}
