% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-1RM.R
\name{update_1RM}
\alias{update_1RM}
\title{Update 1RM}
\usage{
update_1RM(
  LEV_profile,
  what = "visit 1RM",
  athlete = NULL,
  visit = NULL,
  updated_1RM = NULL
)
}
\arguments{
\item{LEV_profile}{\code{LEV_profile} object, returned by \code{\link{create_visits}} function}

\item{what}{Character string indicating whether \code{visit 1RM} or \code{prescription 1RM} should be updated}

\item{athlete}{Vector of athletes to be updated. Default is \code{NULL}}

\item{visit}{Vector of visits to be updated. Default is \code{NULL}}

\item{updated_1RM}{Vector of 1RM values to be applied. Default is \code{NULL}}
}
\description{
Function that updates 1RM in the LEV profile object.
}
\examples{
test <- create_athletes(1) \%>\%
  create_visits(1:5) \%>\%
  create_visit_1RM()

test_df <- test \%>\%
  as.data.frame()

updated_df <- test \%>\%
  update_1RM(what = "visit 1RM", athlete = "Athlete 1", visit = 3, updated_1RM = 200) \%>\%
  as.data.frame()

test_df$visit_1RM
updated_df$visit_1RM
}
