% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-athletes.R
\name{create_athletes}
\alias{create_athletes}
\title{Create Random Athletes}
\usage{
create_athletes(
  n = 20,
  V0_min = 1.5,
  V0_max = 2.5,
  V0_rep_drop_additive_min = 0,
  V0_rep_drop_additive_max = 0,
  V0_rep_drop_multiplicative_min = 0,
  V0_rep_drop_multiplicative_max = 0,
  V0_fatigue_additive_min = 0,
  V0_fatigue_additive_max = 0,
  V0_fatigue_multiplicative_min = 0,
  V0_fatigue_multiplicative_max = 0,
  V0_visit_change_additive_min = 0,
  V0_visit_change_additive_max = 0,
  V0_visit_change_multiplicative_min = -0.05,
  V0_visit_change_multiplicative_max = 0.05,
  V0_visit_random_additive_min = 0,
  V0_visit_random_additive_max = 0,
  V0_visit_random_multiplicative_min = 0.01,
  V0_visit_random_multiplicative_max = 0.05,
  L0_min = 180,
  L0_max = 280,
  L0_rep_drop_additive_min = 0,
  L0_rep_drop_additive_max = 0,
  L0_rep_drop_multiplicative_min = 0.3,
  L0_rep_drop_multiplicative_max = 0.6,
  L0_fatigue_additive_min = 0,
  L0_fatigue_additive_max = 0,
  L0_fatigue_multiplicative_min = 0,
  L0_fatigue_multiplicative_max = 0,
  L0_visit_change_additive_min = 0,
  L0_visit_change_additive_max = 0,
  L0_visit_change_multiplicative_min = -0.05,
  L0_visit_change_multiplicative_max = 0.05,
  L0_visit_random_additive_min = 0,
  L0_visit_random_additive_max = 0,
  L0_visit_random_multiplicative_min = 0.01,
  L0_visit_random_multiplicative_max = 0.05,
  v1RM_min = 0.2,
  v1RM_max = 0.4,
  v1RM_random_additive_min = 0,
  v1RM_random_additive_max = 0,
  v1RM_random_multiplicative_min = 0,
  v1RM_random_multiplicative_max = 0.05,
  est_RIR_systematic_additive_min = -2,
  est_RIR_systematic_additive_max = 2,
  est_RIR_systematic_multiplicative_min = 0,
  est_RIR_systematic_multiplicative_max = 0,
  est_RIR_random_additive_min = 0,
  est_RIR_random_additive_max = 0,
  est_RIR_random_multiplicative_min = 0,
  est_RIR_random_multiplicative_max = 0.25,
  biological_variation_additive_min = 0,
  biological_variation_additive_max = 0,
  biological_variation_multiplicative_min = 0.02,
  biological_variation_multiplicative_max = 0.075,
  instrumentation_noise_additive_min = 0,
  instrumentation_noise_additive_max = 0,
  instrumentation_noise_multiplicative_min = 0,
  instrumentation_noise_multiplicative_max = 0,
  load_increment = 2.5
)
}
\arguments{
\item{n}{Number of athletes}

\item{V0_min}{Numeric}

\item{V0_max}{Numeric}

\item{V0_rep_drop_additive_min}{Numeric}

\item{V0_rep_drop_additive_max}{Numeric}

\item{V0_rep_drop_multiplicative_min}{Numeric}

\item{V0_rep_drop_multiplicative_max}{Numeric}

\item{V0_fatigue_additive_min}{Numeric}

\item{V0_fatigue_additive_max}{Numeric}

\item{V0_fatigue_multiplicative_min}{Numeric}

\item{V0_fatigue_multiplicative_max}{Numeric}

\item{V0_visit_change_additive_min}{Numeric}

\item{V0_visit_change_additive_max}{Numeric}

\item{V0_visit_change_multiplicative_min}{Numeric}

\item{V0_visit_change_multiplicative_max}{Numeric}

\item{V0_visit_random_additive_min}{Numeric}

\item{V0_visit_random_additive_max}{Numeric}

\item{V0_visit_random_multiplicative_min}{Numeric}

\item{V0_visit_random_multiplicative_max}{Numeric}

\item{L0_min}{Numeric}

\item{L0_max}{Numeric}

\item{L0_rep_drop_additive_min}{Numeric}

\item{L0_rep_drop_additive_max}{Numeric}

\item{L0_rep_drop_multiplicative_min}{Numeric}

\item{L0_rep_drop_multiplicative_max}{Numeric}

\item{L0_fatigue_additive_min}{Numeric}

\item{L0_fatigue_additive_max}{Numeric}

\item{L0_fatigue_multiplicative_min}{Numeric}

\item{L0_fatigue_multiplicative_max}{Numeric}

\item{L0_visit_change_additive_min}{Numeric}

\item{L0_visit_change_additive_max}{Numeric}

\item{L0_visit_change_multiplicative_min}{Numeric}

\item{L0_visit_change_multiplicative_max}{Numeric}

\item{L0_visit_random_additive_min}{Numeric}

\item{L0_visit_random_additive_max}{Numeric}

\item{L0_visit_random_multiplicative_min}{Numeric}

\item{L0_visit_random_multiplicative_max}{Numeric}

\item{v1RM_min}{Numeric}

\item{v1RM_max}{Numeric}

\item{v1RM_random_additive_min}{Numeric}

\item{v1RM_random_additive_max}{Numeric}

\item{v1RM_random_multiplicative_min}{Numeric}

\item{v1RM_random_multiplicative_max}{Numeric}

\item{est_RIR_systematic_additive_min}{Numeric}

\item{est_RIR_systematic_additive_max}{Numeric}

\item{est_RIR_systematic_multiplicative_min}{Numeric}

\item{est_RIR_systematic_multiplicative_max}{Numeric}

\item{est_RIR_random_additive_min}{Numeric}

\item{est_RIR_random_additive_max}{Numeric}

\item{est_RIR_random_multiplicative_min}{Numeric}

\item{est_RIR_random_multiplicative_max}{Numeric}

\item{biological_variation_additive_min}{Numeric}

\item{biological_variation_additive_max}{Numeric}

\item{biological_variation_multiplicative_min}{Numeric}

\item{biological_variation_multiplicative_max}{Numeric}

\item{instrumentation_noise_additive_min}{Numeric}

\item{instrumentation_noise_additive_max}{Numeric}

\item{instrumentation_noise_multiplicative_min}{Numeric}

\item{instrumentation_noise_multiplicative_max}{Numeric}

\item{load_increment}{Numeric}
}
\value{
Function returns \code{LV_profile} object
}
\description{
This function create random \code{n} athletes with random \code{LEV_profile} parameters. These
parameters are generated using uniform random number between \code{min} and \code{max} parameters.
LEV profiles can be created either using \code{create_athletes} or \code{\link{create_profiles}}
function. For extensive description of parameters see \code{\link{create_profiles}} function.
}
\examples{
# Create random athletes/profiles
set.seed(1667)

sets <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(90, 110, 130),
    load_type = "absolute"
  )

print(sets)

# Extract data frame
LEV_data <- as.data.frame(sets)
# Or
# LEV_data <- coef(sets)

plot(sets)
plot(sets, athletes = "Athlete 1")
plot(sets, athletes = "Athlete 1", reps = 1)
plot(sets, athletes = "Athlete 1", x_var = "RIR")

# Another way to create LEV profiles
sets <- create_profiles(athlete = c("Mladen", "Ivan"), L0 = c(200, 180)) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(100, 120, 140), load_type = "absolute")

plot(sets)
plot(sets, facet = NULL, x_var = "load")
plot(sets, visits = 1, x_var = "RIR")
}
\seealso{
\code{\link[=create_profiles]{create_profiles()}} function generates LEV profiles, but provide more control of the parameters
}
