% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-profiles.R
\name{create_profiles}
\alias{create_profiles}
\title{Create Load-Exertion-Velocity Profiles}
\usage{
create_profiles(
  athletes = NA,
  V0 = 1.8,
  V0_rep_drop = 0,
  V0_visit_change = 0,
  V0_visit_change_multiplicative = FALSE,
  V0_visit_random = 0,
  V0_visit_random_multiplicative = FALSE,
  L0 = 180,
  L0_rep_drop = 0.04,
  L0_visit_change = 0,
  L0_visit_change_multiplicative = FALSE,
  L0_visit_random = 0,
  L0_visit_random_multiplicative = FALSE,
  v1RM = 0.2,
  v1RM_random = 0,
  v1RM_random_multiplicative = FALSE,
  biological_variation = 0,
  biological_variation_multiplicative = TRUE,
  instrumentation_noise = 0,
  instrumentation_noise_multiplicative = FALSE,
  load_increment = 2.5
)
}
\arguments{
\item{athletes}{Numeric or character vector indicating different athletes. Can also be names}

\item{V0}{Highest velocity when load equals zero}

\item{V0_rep_drop}{Number indicating proportion of \code{V0} dropping
when multiple repetitions are performed. This parameter models the
exertion-velocity characteristics}

\item{V0_visit_change}{Number indicating systematic change in \code{V0} per
each visit to the laboratory/gym}

\item{V0_visit_change_multiplicative}{Should multiplicative change be used? Default
is \code{FALSE}}

\item{V0_visit_random}{Number indicating SD of random change in \code{V0} for
each visit to the laboratory/gym}

\item{V0_visit_random_multiplicative}{Should multiplicative random error be used? Default
is \code{FALSE}}

\item{L0}{Highest load when velocity equals zero}

\item{L0_rep_drop}{Number indicating proportion of \code{L0} dropping
when multiple repetitions are performed. This parameter models the
exertion-velocity characteristics}

\item{L0_visit_change}{Number indicating systematic change in \code{L0} per
each visit to the laboratory/gym}

\item{L0_visit_change_multiplicative}{Should multiplicative change be used? Default
is \code{FALSE}}

\item{L0_visit_random}{Number indicating SD of random change in \code{L0} for
each visit to the laboratory/gym}

\item{L0_visit_random_multiplicative}{Should multiplicative random error be used? Default
is \code{FALSE}}

\item{v1RM}{Velocity at 1RM, or Minimum Velocity Threshold}

\item{v1RM_random}{Number indicating SD of random change in \code{v1RM} for
each visit to the laboratory/gym}

\item{v1RM_random_multiplicative}{Should multiplicative random error be used? Default
is \code{FALSE}}

\item{biological_variation}{Number indicating SD of the random biological variation in velocity
when performing repetitions}

\item{biological_variation_multiplicative}{Should multiplicative random error be used? Default
is \code{TRUE}}

\item{instrumentation_noise}{Number indicating SD of the random instrumentation noise when
measuring repetition velocity}

\item{instrumentation_noise_multiplicative}{Should multiplicative random error be used? Default
is \code{FALSE}}

\item{load_increment}{What should be the minimal jump in load? Default is 2.5, common load jump
in kg in resistance training}
}
\value{
Function returns \code{LV_profile} object
}
\description{
Function creates Load-Exertion-Velocity profiles, which is needed for other functions in the package.
Multiple profiles can be created
}
\examples{
# Create random athletes/profiles
set.seed(1667)

sets <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(90, 110, 130),
    load_type = "absolute"
  )

print(sets)

# Extract data frame
LEV_data <- as.data.frame(sets)
# Or
# LEV_data <- coef(sets)

plot(sets)
plot(sets, athletes = "Athlete 1")
plot(sets, athletes = "Athlete 1", reps = 1)
plot(sets, athletes = "Athlete 1", x_var = "RIR")

# Another way to create LEV profiles
sets <- create_profiles(athletes = c("Mladen", "Ivan"), L0 = c(200, 180)) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(100, 120, 140), load_type = "absolute")

plot(sets)
plot(sets, facet = NULL, x_var = "load")
plot(sets, visits = 1, x_var = "RIR")
}
\seealso{
\code{\link[=create_athletes]{create_athletes()}} is another option to create LEV profiles
}
