% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sets.R
\name{create_sets}
\alias{create_sets}
\title{Create Sets}
\usage{
create_sets(
  LEV_profile,
  load,
  reps = rep(NA, length(load)),
  target_est_RIR = rep(NA, length(load)),
  target_velocity = rep(NA, length(load)),
  target_VL = rep(NA, length(load)),
  target_est_MNR = rep(NA, length(load)),
  load_type = "absolute",
  max_reps = 100,
  failed_reps = FALSE,
  failed_sets = FALSE,
  use_true_velocity = FALSE,
  inter_set_fatigue = TRUE,
  incremental_est_RIR = TRUE
)
}
\arguments{
\item{LEV_profile}{\code{LEV_profile} object, returned by \code{\link{create_visits}} function}

\item{load}{Numeric vector. Loads are either absolute weight or percentages. See also \code{load_type}}

\item{reps}{Target number of reps. Default is equal to \code{NA}, implying max reps}

\item{target_est_RIR}{Target \code{est_RIR}. Set is stopped when reaching this quality threshold. Default is \code{NA}}

\item{target_velocity}{Target measured velocity. Set is stopped when reaching this quality threshold. Default is \code{NA}}

\item{target_VL}{Target velocity loss. Set is stopped when reaching this quality threshold. Default is \code{NA}}

\item{target_est_MNR}{Target \code{est_MNR}. Set is stopped when reaching this quality threshold. Default is \code{NA}}

\item{load_type}{Type of load calculation. Can be either 'absolute' (default), 'profile 1RM', 'L0',
'visit 1RM', or 'prescription 1RM'}

\item{max_reps}{How many maximum reps to generate to search for failure? Default is 100}

\item{failed_reps}{Should failed-reps be included in the output? Default is \code{FALSE}}

\item{failed_sets}{Should failed-sets be included in the output? Default is \code{FALSE}}

\item{use_true_velocity}{When estimating failure, should true or biological (default) velocity be used?}

\item{inter_set_fatigue}{Should profile inter-set fatigue parameters be utilized? Default is \code{TRUE}}

\item{incremental_est_RIR}{When calculating \code{est_RIR} for the preceding reps last rep, should
increments be used or instantaneous subjective ratings for each rep? Default is \code{TRUE}}
}
\value{
Object \code{LEV_sets}
}
\description{
Function creates sets, using provided \code{LEV_profile}, \code{loads}, and \code{trials}
}
\examples{
# Create random athletes/profiles
set.seed(1667)

sets <- create_athletes(5) \%>\%
  create_visits(1:3) \%>\%
  create_sets(
    load = c(90, 110, 130),
    load_type = "absolute"
  )

print(sets)

# Extract data frame
LEV_data <- as.data.frame(sets)
# Or
# LEV_data <- coef(sets)

plot(sets)
plot(sets, athletes = "Athlete 1")
plot(sets, athletes = "Athlete 1", reps = 1)
plot(sets, athletes = "Athlete 1", x_var = "RIR")

# Another way to create LEV profiles
sets <- create_profiles(athlete = c("Mladen", "Ivan"), L0 = c(200, 180)) \%>\%
  create_visits(1) \%>\%
  create_sets(load = c(100, 120, 140), load_type = "absolute")

plot(sets)
plot(sets, facet = NULL, x_var = "load")
plot(sets, visits = 1, x_var = "RIR")
}
