% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tests.R
\name{create_tests}
\alias{create_tests}
\title{Create tests}
\usage{
create_tests(
  LEV_profile = create_profiles(),
  load_1RM = seq(0.6, 1.2, by = 0.025),
  load_LV = c(0.5, 0.6, 0.7, 0.8, 0.85, 0.9, 0.95, 0.975, 1),
  load_RTF = c(0.9, 0.8, 0.7),
  max_reps = 100,
  failed_reps = FALSE,
  failed_sets = FALSE,
  use_true_velocity = FALSE
)
}
\arguments{
\item{LEV_profile}{\code{LEV_profile} object, returned by \code{\link{create_visits}} function}

\item{load_1RM}{Percentages used to estimate visit 1RM. Percentage are used to calculate loads using theoretical visit profile 1RM}

\item{load_LV}{Percentages used to estimate LV profile}

\item{load_RTF}{Percentages used to estimate RTF profile}

\item{max_reps}{How many maximum reps to generate to search for failure? Default is 100}

\item{failed_reps}{Should failed-reps be included in the output? Default is \code{FALSE}}

\item{failed_sets}{Should failed-sets be included in the output? Default is \code{FALSE}}

\item{use_true_velocity}{When estimating failure, should true or biological (default) velocity be used?}
}
\value{
Object \code{LEV_sets}
}
\description{
Estimate individual 1RM, create Load-Velocity and Reps-To-Failure profiles.
The set fatigue effects, set in the individual profiles, are disregarded when
creating LV and RTF profiles
}
\examples{
test_sets <- create_profiles(athletes = 1:2, L0_visit_random = 10, L0_fatigue = -10) \%>\%
  create_visits(1) \%>\%
  create_tests() \%>\%
  create_summary()

plot(test_sets, sets = "LV")
plot(test_sets, sets = "RTF")
}
