% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-S3.R
\name{plot.LEV_summary}
\alias{plot.LEV_summary}
\title{S3 method for plotting LEV Sets Summaries}
\usage{
\method{plot}{LEV_summary}(x, type = "athletes", ...)
}
\arguments{
\item{x}{Object of class \code{LEV_sets}}

\item{type}{Type of plot. Default is "athletes" which creates facets
for each athlete. Another option is "pooled", which doesn't create
facets for athletes by default. One can use \code{facet} parameter
to create facets using another variable. See Details}

\item{...}{Extra arguments. See Details}
}
\description{
S3 method for plotting LEV Sets Summaries
}
\details{
S3 method for plotting LEV Sets Summaries provides plenty of
flexibility and control by using the \code{...} parameters.
User can filter variables, provides colors and facets.
Here are the extra parameters your can utilize:
\describe{
\item{athletes}{Althetes to keep in}
\item{visits}{Visits to keep in}
\item{sets}{Sets to keep in}
\item{loads}{Loads to keep in}
\item{nRMs}{N-Repetition-Maximum loads to keep in}
\item{est_nRMs}{Estimated N-Repetition-Maximum loads to keep in}
\item{sets_to_failure}{Sets-To-Failure indicator (TRUE/FALSE) to keep in}
\item{target_repetitions}{Target repetitions to keep in}
\item{repetitions_done}{Repetitions done to keep in}
\item{RIRs}{Reps-In-Reserve to keep in}
\item{est_RIRs}{Estimated Reps-In-Reserve to keep in}
\item{color}{What variable should be used to coloring groups? Default is none}
\item{facet}{Variable to be used for plot faceting. Athletes are already faceted,
and this parameter allows for adding extra level. Default is 'visit'}
\item{label_size}{Label size}
}

Please check the examples section for few examples
}
\examples{
set.seed(1667)
sets <- create_athletes(2) \%>\%
  create_visits(1) \%>\%
  create_visit_1RM() \%>\%
  create_sets(load = c(0.7, 0.8, 0.9), load_type = "visit 1RM") \%>\%
  create_summary()

sets
plot(sets)
}
